+++ CHANGELOG.md
@@ -8,6 +8,7 @@ script to proceed.
 - New database migration engine.
 - Removed parameter of `temboard --debug` CLI option.
 - Move `temboard-migratedb` as `temboard migratedb` subcommand.
+- New subcommand `temboard runtask` for debugging.
 
 
 ## [7.11] - Unreleased